Revert "start of #10 - Start of C-64 & Joystick support"
