+++ board.h
@@ -3,13 +3,10 @@
 
 // exported functions
 extern void getboard ();
-extern void updatescore ();
-extern void updateboard ();
-
-#ifdef __CX16__
 extern void stashboard ();
 extern void fetchboard ();
-#endif
+extern void updatescore ();
+extern void updateboard ();
 
 // exported vars
 extern char board[16][16];
