+++ boardedit.h
@@ -4,6 +4,5 @@
 // exported functions
 extern void markcol (unsigned char mx,unsigned char my);
 extern void markrow (unsigned char mx,unsigned char my);
-extern void loadboard ();
-extern void saveboard ();
+
 #endif
