+++ button.c
@@ -6,15 +6,10 @@
 #include "board.h"
 #include "boardedit.h"
 
-#ifdef __CX16__
-  #define CAN_EDIT
-#endif
-
 bool global_editmode;
 bool global_newrandboard;
 
 void mbutton (unsigned char mx, unsigned char my) {
-	#ifdef CAN_EDIT
 	if (global_editmode) {
 		// 'load' button
 		if ((mx >= 31 && mx <= 37) && (my >= 11 && my <= 13)) {
@@ -39,7 +34,6 @@ void mbutton (unsigned char mx, unsigned char my) {
 			markrow (mx,my);
 		}
 	}
-	#endif
 	// clicked on game 'reset' button
 	if ((mx >= 31 && mx <= 37) && (my >= 5 && my <= 7)) {
 		global_newrandboard = true;
@@ -62,12 +56,10 @@ void mbutton (unsigned char mx, unsigned char my) {
 			updateboard ();
 		}
 	}
-	#ifdef __CX16__
 	// clicked on the 'replay' button
 	else if ((mx >= 30 && mx <= 37) && (my >= 15 && my <= 17)) {
 		fetchboard ();
 		global_newrandboard = false;
 		updateboard ();
 	}
-	#endif
 }
