--- c64/joy.h
@@ -1,10 +1,0 @@
-#ifndef JOY_H
-#define JOY_H
-
-// exported variables
-extern short jx,jy;
-
-// exported functions
-extern void joyinfo ();
-
-#endif
