+++ queue.c
@@ -11,17 +11,15 @@ unsigned char dirtable[4][2] = {
 	{ 107 , 106 }
 };
 
-#ifdef __CX16__
-// these numbers are the screen display codes
-// corresponding PETSCII codes are 100-103
+// screen display codes
 unsigned char newchars[8] = { 68,69,70,71 };
 
 // circle halfs
 unsigned char ch[4][8] = {
-	{ 0,0,0,60,126,231,195,195 },   // top
-	{ 224,240,56,24,24,56,240,224 }, // right
-	{ 195,195,231,126,60,0,0,0 },   // bottom
-	{ 7,15,28,24,24,28,15,7 }      // left
+	{ 0,0,0,126,255,231,195,195 },   // top
+	{ 240,248,56,24,24,56,248,240 }, // right
+	{ 195,195,231,255,126,0,0,0 },   // bottom
+	{ 15,31,28,24,24,28,31,15 }      // left
 };
 
 void setupchars () {
@@ -35,7 +33,6 @@ void setupchars () {
 		}
 	}
 }
-#endif
 
 Queue * createQueue (unsigned int maxElements) {
 	Queue *Q;
@@ -187,14 +184,12 @@ void processQ (Queue *Q) {
 		}
 
 		textcolor (COLOR_WHITE);
-		#ifdef __CX16__
 		cputcxy (mx,my,hc);
-		#endif
-		vsyncw (3);
+		vsyncw (2);
 		cputcxy (mx,my,nc);
 
 		Deqeue (Q);
 
-		//vsyncw (1);
+		vsyncw (1);
 	}
 }
